backend/requirements.txt
flask
flask-cors
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai
